Separate common and raw
