+++ lib/ansible/module_utils/k8s/lookup.py
@@ -21,12 +21,11 @@ from __future__ import absolute_import, division, print_function
 import json
 import os
 
-from ansible.module_utils.k8s.common import DateTimeEncoder, remove_secret_data, to_snake
+from ansible.module_utils.k8s.common import OpenShiftAnsibleModuleMixin, DateTimeEncoder, remove_secret_data, to_snake
 from ansible.module_utils.k8s.helper import AUTH_ARG_SPEC
 
 try:
     from openshift.helper.kubernetes import KubernetesObjectHelper
-    from openshift.helper.openshift import OpenShiftObjectHelper
     from openshift.helper.exceptions import KubernetesException
     HAS_K8S_MODULE_HELPER = True
 except ImportError as exc:
@@ -88,7 +87,7 @@ class KubernetesLookup(object):
             )
 
         self.kind = to_snake(self.kind)
-        self.helper = self.get_helper()
+        self.helper = self.get_helper(self.api_version, self.kind)
 
         for arg in AUTH_ARG_SPEC:
             self.connection[arg] = kwargs.get(arg)
@@ -110,10 +109,10 @@ class KubernetesLookup(object):
         with open('loggit.txt', 'a') as f:
             f.write(msg + '\n')
 
-    def get_helper(self):
+    def get_helper(self, api_version, kind):
         try:
-            helper = KubernetesObjectHelper(api_version=self.api_version, kind=self.kind, debug=False)
-            helper.get_model(self.api_version, self.kind)
+            helper = KubernetesObjectHelper(api_version=api_version, kind=kind, debug=False)
+            helper.get_model(api_version, kind)
             return helper
         except KubernetesException as exc:
             raise Exception("Error initializing helper: {0}".format(exc.message))
@@ -208,12 +207,5 @@ class KubernetesLookup(object):
         return response
 
 
-class OpenShiftLookup(KubernetesLookup):
-
-    def get_helper(self):
-        try:
-            helper = OpenShiftObjectHelper(api_version=self.api_version, kind=self.kind, debug=False)
-            helper.get_model(self.api_version, self.kind)
-            return helper
-        except KubernetesException as exc:
-            raise Exception("Error initializing helper: {0}".format(exc.message))
+class OpenShiftLookup(OpenShiftAnsibleModuleMixin, KubernetesLookup):
+    pass
