+++ test/sanity/pylint/ignore.txt
@@ -6,8 +6,6 @@ hacking/cherrypick.py ansible-format-automatic-specification
 hacking/metadata-tool.py ansible-format-automatic-specification
 lib/ansible/cli/adhoc.py syntax-error 3.7
 lib/ansible/module_utils/dimensiondata.py ansible-format-automatic-specification
-lib/ansible/module_utils/k8s/common.py ansible-format-automatic-specification
-lib/ansible/module_utils/k8s/scale.py ansible-format-automatic-specification
 lib/ansible/module_utils/network/aci/aci.py ansible-format-automatic-specification
 lib/ansible/module_utils/network/iosxr/iosxr.py ansible-format-automatic-specification
 lib/ansible/module_utils/ovirt.py ansible-format-automatic-specification
